% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{COVIDmodel}
\alias{COVIDmodel}
\title{Runs the conisi COVID-19 model}
\usage{
COVIDmodel(
  parm_table,
  pop_size,
  num_days,
  pop_prop,
  contact_matrix,
  solver_method = "euler",
  param_method = "linear-interpolation"
)
}
\arguments{
\item{parm_table}{A data frame containing the time-varying parameters (in long format).}

\item{pop_size}{Integer The size of the population being modelled.}

\item{num_days}{Integer How many days to run the simulation for.}

\item{pop_prop}{Vector This vector contains population proportions for sub-populations}

\item{contact_matrix}{Vector The entries of the mixing matrix}

\item{solver_method}{String The name of the solver method to use with deSolve. Default is euler.}

\item{param_method}{String How the params should be prepared. Options are "step" and "linear-interpolation".}
}
\value{
A data frame with the values of the various compartments over time
}
\description{
Solves the system of differential equations that make up the conisi model
using the parameters and arguments provided.
}
