% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{COVIDmodel_run_and_mutate_many}
\alias{COVIDmodel_run_and_mutate_many}
\title{Runs the conisi COVID-19 model on multiple parameter sets at once and applies mutate_model to the result}
\usage{
COVIDmodel_run_and_mutate_many(parm_table, pop_size, num_days)
}
\arguments{
\item{parm_table}{A data frame containing the time-varying parameters (in long format) for multiple experiments.}

\item{pop_size}{Integer The size of the population being modelled.}

\item{num_days}{Integer How many days to run the simulation for.}
}
\value{
A data frame with the values of the various compartments
}
\description{
Runs the conisi COVID-19 model on multiple parameter sets at once and applies mutate_model to the result
}
