% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_parameters.R
\name{load_parameters}
\alias{load_parameters}
\title{Loads simulation parameters from a file}
\usage{
load_parameters(file_path, sim_name = NULL, start_date = NULL)
}
\arguments{
\item{file_path}{String The path to the file to be loaded}

\item{sim_name}{String Optional The name of the simulation to be extracted from the file}

\item{start_date}{Date Must be provided if the file contains start_date instead of start_time.
It is the date that will be considered time zero.}
}
\description{
This function reads parameters from a file and reformats them so they are ready to be used in the model
}
