% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{modelrmse}
\alias{modelrmse}
\title{Calculates model RMSE}
\usage{
modelrmse(
  modelOutput,
  start_date,
  target_data,
  weights = c(1, 0.5),
  under_report_factor = 1
)
}
\arguments{
\item{modelOutput}{results from the model}

\item{start_date}{date The date which the first model time-point is being compared to.}

\item{target_data}{Data frame The observed data that the model is targeted to match.}

\item{weights}{Vector Vector of length 2 containing weights indicating the relative weighting to give cases and deaths.
For example  c(1.0, 0.5)}

\item{under_report_factor}{fraction Estimated fraction of deaths that get reported as covid-19 deaths. This is deprecated and no longer used.
The estimated_deaths in the variable "cum_est_deaths" allows for adjusting the death reporting rates.}
}
\description{
Calculates model RMSE
}
