% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{prep_params}
\alias{prep_params}
\title{prep_params produces a timeseries of parameter values using the specified method}
\usage{
prep_params(parm_table, times, param_method = "linear-interpolation")
}
\arguments{
\item{parm_table}{data.frame The params as loaded using load_parameters()}

\item{times}{vector of integer values for which to generate param values}

\item{param_method}{char The method to use, one of "step", "linear-interpolation".}
}
\description{
prep_params produces a timeseries of parameter values using the specified method
}
