% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{COVIDmodel_run_and_mutate}
\alias{COVIDmodel_run_and_mutate}
\title{Runs the conisi COVID-19 model and applies mutate_model to the result}
\usage{
COVIDmodel_run_and_mutate(
  parm_table,
  pop_size,
  num_days,
  pop_prop,
  contact_matrix,
  start_date = NULL,
  report_lag = 0
)
}
\arguments{
\item{parm_table}{A data frame containing the time-varying parameters (in long format).}

\item{pop_size}{Integer The size of the population being modelled.}

\item{num_days}{Integer How many days to run the simulation for.}

\item{pop_prop}{Vector This vector contains population proportions for sub-populations}

\item{contact_matrix}{Vector The entries of the mixing matrix}

\item{start_date}{Date This is the start date for the local epidemic and it is used for adding a date column to the output.}

\item{report_lag}{Integer This is the number of days we assume pass between the infection and when it is first reported.}
}
\value{
A data frame with the values of the various compartments, parameters and other statistics over time
}
\description{
Runs the conisi COVID-19 model and applies mutate_model to the result
}
