% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{modelrmse}
\alias{modelrmse}
\title{Calculates model RMSE}
\usage{
modelrmse(
  modelOutput = modelOutput,
  local_epi_start_date_0 = local_epi_start_date,
  local_obs_start_date_0 = local_obs_start_date,
  data = region_data,
  report_lag_0 = report_lag,
  weights = region_weights,
  under_report_factor = 1
)
}
\arguments{
\item{modelOutput}{results from the model}

\item{local_epi_start_date_0}{The date on which the model results start}

\item{local_obs_start_date_0}{The data on which the observed data start}

\item{data}{Data frame The observed data}

\item{report_lag_0}{integer The lag between model data and observed data}

\item{weights}{Vector Leght 2 vector containing weights indicating the relative weighting to give cases and deaths}

\item{under_report_factor}{fraction Estimated fraction of deaths that get reported as covid-19 deaths}
}
\description{
Calculates model RMSE
}
